requirements.txt
tensorflow
numpy
opencv-python
flask
python-socketio
python-dotenv
eventlet
Cython
